FlaskApp/requirements.txt
click==8.1.3
Flask==2.1.2
Flask-JWT-Extended==4.4.0
Flask-Login==0.6.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Pillow==9.1.1
PyJWT==2.4.0
SQLAlchemy==1.4.36
Werkzeug==2.1.2
zipp==3.8.0
